demo : initSharedTexture
